+++ src/Application/View/Participant/Export/ParticipantView.php
@@ -61,6 +61,9 @@ class ParticipantView
     /** @var int */
     public $scheduledMeetings;
 
+    /** @var int */
+    public $chatSessionsCallVisio;
+
     public function __construct(
         int $sheetId,
         string $typeTitle,
@@ -80,7 +83,8 @@ class ParticipantView
         int $viewedSheets,
         int $clickedElements,
         int $requestedMeetings,
-        int $scheduledMeetings
+        int $scheduledMeetings,
+        int $chatSessionsCallVisio
     ) {
         $this->sheetId = $sheetId;
         $this->typeTitle = $typeTitle;
@@ -101,5 +105,6 @@ class ParticipantView
         $this->clickedElements = $clickedElements;
         $this->requestedMeetings = $requestedMeetings;
         $this->scheduledMeetings = $scheduledMeetings;
+        $this->chatSessionsCallVisio = $chatSessionsCallVisio;
     }
 }