+++ tests/Application/Query/Agenda/AgendaViewQueryHandlerTest.php
@@ -1,5 +1,13 @@
 <?php
 
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
 namespace Proximum\Vimeet\Tests\Application\Query\Agenda;
 
 use PHPUnit\Framework\TestCase;
@@ -432,8 +440,6 @@ class AgendaViewQueryHandlerTest extends TestCase
             [
                 111 => new MeetingView(
                     111,
-                    42,
-                    24,
                     'Sheet 1',
                     1243,
                     [new SheetMetView('Sheet other', false)],
@@ -454,8 +460,6 @@ class AgendaViewQueryHandlerTest extends TestCase
                 ),
                 222 => new MeetingView(
                     111,
-                    43,
-                    25,
                     'Sheet 1',
                     14883,
                     [new SheetMetView('Another sheet', false)],