+++ tests/Application/Query/Participant/Export/ParticipantViewQueryHandlerTest.php
@@ -16,6 +16,7 @@ use Proximum\Vimeet\Domain\Model\ProductAttributedToParticipant;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\User;
+use Proximum\Vimeet\Domain\Repository\ChatSessionRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\HappeningParticipationRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\HappeningRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\MeetingRepositoryInterface;
@@ -55,6 +56,9 @@ class ParticipantViewQueryHandlerTest extends TestCase
 
     /** @var ObjectProphecy */
     private $meetingRepository;
+
+    /** @var ObjectProphecy */
+    private $chatSessionRepository;
 
     /** @var ObjectProphecy */
     private $scanRepository;
@@ -105,6 +109,7 @@ class ParticipantViewQueryHandlerTest extends TestCase
         $this->sheetRepository = $this->prophesize(SheetRepositoryInterface::class);
         $this->requestRepository = $this->prophesize(RequestRepositoryInterface::class);
         $this->meetingRepository = $this->prophesize(MeetingRepositoryInterface::class);
+        $this->chatSessionRepository = $this->prophesize(ChatSessionRepositoryInterface::class);
 
         $this->participant = $this->prophesize(Participant::class);
         $this->event = $this->prophesize(Event::class);
@@ -265,6 +270,10 @@ class ParticipantViewQueryHandlerTest extends TestCase
             ->shouldBeCalled()
             ->willReturn(3);
 
+        $this->chatSessionRepository->countCallVisioByEventAndUser($this->event->reveal(), $this->user->reveal())
+            ->shouldBeCalled()
+            ->willReturn(7);
+
         $handler = new ParticipantViewQueryHandler(
             $this->happeningParticipationRepository->reveal(),
             $this->templateDataFactory->reveal(),
@@ -275,7 +284,8 @@ class ParticipantViewQueryHandlerTest extends TestCase
             $happeningRepository->reveal(),
             $this->sheetRepository->reveal(),
             $this->requestRepository->reveal(),
-            $this->meetingRepository->reveal()
+            $this->meetingRepository->reveal(),
+            $this->chatSessionRepository->reveal()
         );
 
         $result = $handler->handle($query);
@@ -311,7 +321,8 @@ class ParticipantViewQueryHandlerTest extends TestCase
             42,
             24,
             5,
-            3
+            3,
+            7
         );
 
         $this->assertEquals($expected, $result);